Misc improvements

+++ system/bin/manage_fonts.sh
@@ -1,11 +1,12 @@
-#!/data/adb/modules/fontrevival/tools/busybox ash
-# shellcheck shell=ash
+# shellcheck shell=bash
 # shellcheck disable=SC2169
 # shellcheck disable=SC2034
 # shellcheck disable=SC2183
 # shellcheck disable=SC2145
 # shellcheck disable=SC2155
 # shellcheck disable=SC2059
+shopt -s checkwinsize
+resize
 G='\e[01;32m'
 R='\e[01;31m'
 Y='\e[01;33m'
@@ -17,55 +18,16 @@ W='\e[01;37m'
 BGBL='\e[1;30;47m'
 N='\e[0m'
 # shellcheck disable=SC2154
+if test "$(id -u)" -ne 0; then
+	echo -e "Please run this script as root!"
+	exit 1
+fi
 if test -n "${ANDROID_SOCKET_adbd}"; then
-    echo -e "Please run this in a temrinal emulator on device!"
+    echo -e "Please run this in a terminal emulator on device!"
     exit 1
 fi
 COLUMNS="$(stty size | cut -d" " -f2)"
 div="${Bl}$(printf '%*s' $((COLUMNS * 90 / 100)) '' | tr " " "=")${N}"
-it_failed() {
-    set +euxo pipefail
-    if "$1" -ne "0"; then
-        echo -e "${R}============= ⓧ ERROR ⓧ =============${N}"
-        echo -e "${R}Something bad happened and the script has encountered an issue${N}"
-        echo -e "${R}Make sure you're following instructions and try again!${N}"
-        echo -e "${R}============= ⓧ ERROR ⓧ =============${N}"
-        echo -e "Exiting the script now!"
-    fi
-    exit "$1"
-}
-detect_ext_data() {
-    if touch /sdcard/.rw && rm /sdcard/.rw; then
-        export EXT_DATA="/sdcard/FontManager"
-    elif touch /storage/emulated/0/.rw && rm /storage/emulated/0/.rw; then
-        export EXT_DATA="/storage/emulated/0/FontManager"
-    elif touch /data/media/0/.rw && rm /data/media/0/.rw; then
-        export EXT_DATA="/data/media/0/FontManager"
-    else
-        EXT_DATA='/storage/emulated/0/FontManager'
-        echo -e "⚠ Possible internal storage access issues! Please make sure data is mounted and decrypted."
-        echo -e "⚠ Trying to proceed anyway "
-        sleep 2
-    fi
-}
-detect_ext_data
-if test ! -d "$EXT_DATA"; then
-    mkdir -p "$EXT_DATA" >/dev/null
-fi
-if ! touch "EXT_DATA"/.rw && rm -fr "EXT_DATA"/.rw; then
-    if ! rm -fr "$EXT_DATA" && mktouch "EXT_DATA"/.rw && rm -fr "EXT_DATA"/.rw; then
-        echo -e "⚠ Cannot access internal storage!"
-        it_failed
-    fi
-fi
-mkdir -p "$EXT_DATA"/logs >/dev/null
-mkdir -p "$EXT_DATA"/lists >/dev/null
-MODDIR="/data/adb/modules/fontrevival"
-exec 3>&2 2>"$EXT_DATA"/logs/script.log
-set -x 2
-set -eo pipefail
-trap 'it_failed $?' EXIT
-clear
 do_banner() {
     clear
     echo -e "${B}  _____              _                           ${N}"
@@ -85,37 +47,80 @@ do_banner() {
 do_banner
 echo -e "$div"
 echo -e "${G}Loading...${N}"
+it_failed() {
+    set +euxo pipefail
+    if "$1" -ne "0"; then
+        echo -e "${R}============= ⓧ ERROR ⓧ =============${N}"
+        echo -e "${R}Something bad happened and the script has encountered an issue${N}"
+        echo -e "${R}Make sure you're following instructions and try again!${N}"
+        echo -e "${R}============= ⓧ ERROR ⓧ =============${N}"
+        echo -e "Exiting the script now!"
+    fi
+    exit "$1"
+}
+if ! $NR; then
+    echo -e "${R}Do not call this script directly! Instead call just 'manage_fonts'${N}"
+    it_failed
+fi
+detect_ext_data() {
+    if touch /sdcard/.rw && rm /sdcard/.rw; then
+        export EXT_DATA="/sdcard/FontManager"
+    elif touch /storage/emulated/0/.rw && rm /storage/emulated/0/.rw; then
+        export EXT_DATA="/storage/emulated/0/FontManager"
+    elif touch /data/media/0/.rw && rm /data/media/0/.rw; then
+        export EXT_DATA="/data/media/0/FontManager"
+    else
+        EXT_DATA='/storage/emulated/0/FontManager'
+        echo -e "⚠ Possible internal storage access issues! Please make sure data is mounted and decrypted."
+        echo -e "⚠ Trying to proceed anyway "
+        sleep 2
+    fi
+}
+detect_ext_data
+if test ! -d "$EXT_DATA"; then
+    mkdir -p "$EXT_DATA" >/dev/null
+fi
+if ! touch "EXT_DATA"/.rw && rm -fr "EXT_DATA"/.rw; then
+    if ! rm -fr "$EXT_DATA" && touch "EXT_DATA"/.rw && rm -fr "EXT_DATA"/.rw; then
+        echo -e "⚠ Cannot access internal storage!"
+        it_failed
+    fi
+fi
+mkdir -p "$EXT_DATA"/logs >/dev/null
+mkdir -p "$EXT_DATA"/lists >/dev/null
+MODDIR="/data/adb/modules/fontrevival"
+exec 3>&2 2>"$EXT_DATA"/logs/script.log
+set -x 2
+set -o pipefail
+trap 'it_failed' EXIT
 no_i() {
     do_banner
     echo -e "${R}No internet access!${N}"
     echo -e "${R}For now this module requires internet access${N}"
     echo -e "${R}Exiting${N}"
     sleep 3
-    it_failed $?
+    it_failed
 }
 e_spinner() {
     PID=$!
     h=0
-    anim='-\|/'
+    anim='◐◓◑◒'
     while [ -d /proc/$PID ]; do
         h=$(((h + 1) % 4))
-        sleep 0.05
+        sleep 0.02
         # shellcheck disable=SC2145,SC2059
-        printf "\r${@} [${anim:$h:1}]"
+        printf "\r${@}[${anim:$h:1}]"
     done
 }
 do_quit() {
     clear
     do_banner
+    echo -e "$div"
     echo -e "${G}Thanks for using Font Manager${N}"
     echo -e "${G}Goodbye${N}"
     sleep 2
     exit 0
 }
-if ! $NR; then
-    echo -e "${R}Do not call this script directly! Instead call just 'manage_fonts'${N}"
-    it_failed $?
-fi
 test_connection() {
     (ping -q -c 2 -W 2 androidacy.com >/dev/null 2>&1) && return 0 || return 1
 }
@@ -145,28 +150,42 @@ font_select() {
         menu_set
     fi
     if ! grep -i "^$a$" "$MODDIR"/lists/fonts-list.txt >/dev/null; then
-        no_i
+         do_banner
+        echo -e "$div"
+        echo -e "${R}ERROR: INVALID SELECTION${N}"
+        sleep 0.5
+        echo -e "${Y}Please try again${N}"
+        sleep 3
+        font _select
     fi
+	do_banner
+	echo -e "$div"
     test_connection &
     e_spinner "${Y}Checking for internet access ${N}"
     if test $? -ne 0; then
-        do_banner
-        echo -e "${R}No internet access!${N}"
-        echo -e "${R}For now this module requires internet access${N}"
-        echo -e "${R}Exiting${N}"
-        sleep 3
-        it_failed $?
+        no_i
     else
-        do_banner
-        curl -kL https://dl.androidacy.com/downloads/fontifier-files/fonts/"$a".zip >"$EXT_DATA"/"$a".zip && sleep 2 &
+        curl -kL https://dl.androidacy.com/downloads/fontifier-files/fonts/"$a".zip >"$EXT_DATA"/"$a".zip && sleep 1 &
         e_spinner "${G}Downloading $a font ${N}"
         sleep 2
-        unzip "$EXT_DATA"/"$a".zip -d "$MODDIR/system/fonts" >/dev/null && sleep 2 &
+        in_f() { 
+        	unzip "$EXT_DATA"/"$a".zip -d "$MODDIR/system/fonts" >/dev/null
+			if test -d /product/fonts; then
+				mkdir -p "$MODDIR"/system/product/fonts
+				cp "$MODDIR"/system/fonts/* "$MODDIR"/system/product/fonts/
+			fi
+			if test -d /system_ext/fonts; then
+				mkdir -p "$MODDIR"/system/system_ext/fonts
+				cp "$MODDIR"/system/fonts/* "$MODDIR"/system/system_ext/fonts/
+			fi
+			sleep 1
+		}
+		in_f &
         e_spinner "${G}Installing $a font ${N}"
         echo -e " "
         echo -e "${G}Install success!${N}"
         echo "$a" >"$MODDIR"/curr-font.txt
-        sleep 1.5
+        sleep 2
     fi
     menu_set
 }
@@ -196,7 +215,6 @@ emoji_select() {
         menu_set
     fi
     if ! grep -i "^$a$" "$MODDIR"/lists/emojis-list.txt >/dev/null; then
-        clear
         do_banner
         echo -e "$div"
         echo -e "${R}ERROR: INVALID SELECTION${N}"
@@ -205,14 +223,15 @@ emoji_select() {
         sleep 3
         emoji_select
     fi
+	do_banner
+	echo -e "$div"
     test_connection &
     e_spinner "${Y}Checking for internet access ${N}"
     if test $? -ne 0; then
         no_i
     else
-        do_banner
         sleep 0.2
-        curl -kL https://dl.androidacy.com/downloads/fontifier-files/emojis/"$a".zip >"$EXT_DATA"/"$a".zip && sleep 2 &
+        curl -kL https://dl.androidacy.com/downloads/fontifier-files/emojis/"$a".zip >"$EXT_DATA"/"$a".zip && sleep 1 &
         e_spinner "${G}Downloading $a emoji ${N}"
         unzip "$EXT_DATA"/"$a".zip -d "$MODDIR/system/fonts" >/dev/null && sleep 2 &
         e_spinner "${G}Installing $a emoji ${N}"
@@ -237,14 +256,13 @@ update_lists() {
             curl -kL https://dl.androidacy.com/downloads/fontifier-files/lists/emojis-list.txt >"$MODDIR"/lists/emojis-list.txt
             sed -i s/[.]zip//gi "$MODDIR"/lists/*
             cp "$MODDIR"/lists/* "$EXT_DATA"/lists
-            sleep 2
+            sleep 2.5
         }
         dl_l &
         e_spinner "${G}Downloading fresh lists ${N}"
-        sleep 1
         echo -e " "
         echo -e "${Y}Lists updated! Returning to menu${N}"
-        sleep 1
+        sleep 2
         clear
         menu_set
     fi
@@ -268,10 +286,13 @@ detect_others() {
 reboot() {
     do_banner
     echo -e "$div"
-    echo -e "${R}Press Ctrl-C to cancel reboot${N}"
-    sleep 5 &
-    e_spinner "${R} Rebooting in five seconds.${N}"
-    setprop sys.powerctl reboot
+    echo -en "${R}Are you sure you want to reboot? [y/N] ${N}"
+    read -r a
+	if test "$a" == "y"; then
+    	setprop sys.powerctl reboot
+	else
+		menu_set
+	fi
 }
 rever_st() {
     do_banner
@@ -282,13 +303,12 @@ rever_st() {
     }
     r_s &
     e_spinner "${G}Reverting to stock fonts ${N}"
-    echo -e "${G}Stock fonts applied! Please reboot.${N}"
+    echo -e "\n${G}Stock fonts applied! Please reboot.${N}"
     sleep 2
     menu_set
 }
 menu_set() {
     while :; do
-        detect_others
         do_banner
         echo -e "$div"
         for i in curr-font curr-emoji; do
@@ -320,4 +340,5 @@ menu_set() {
         esac
     done
 }
+detect_others
 menu_set
